TTDA-29 reorganized app to allow integration test to run
